requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.129.0
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.2.6
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
